Move animations.py into RimSort.models
